Keep keepass file open
This commit leaves the keepasscx file open, enabling faster unlock at
the cost of missing auto-reload when entries do change. This has the
benefit of us not needing to store the password in-memory during the
scripts runtime.

This commit also adds multiple new options with the aim of making this
script run under wayland. These are:

kpx_dotool: wether xdotool, dotool, or any other dotool should be used
kpx_kpx_dotool_prefix: A prefix prepended to the password entry.
kpx_notify: A command to be executed after the password insert.

Whilest I was at it, I also enabled the script to ping the running
daemon if it is executed a second time, making integration as a keybind
easier.

Signed-off-by: RBO <drain+git@rbo.wf>